Make links withe method work everywhere, not only inside turbo frames
