+++ src/core/frames/link_interceptor.ts
@@ -1,5 +1,3 @@
-import { dispatch } from "../../util"
-
 export interface LinkInterceptorDelegate {
   shouldInterceptLinkClick(element: Element, url: string): boolean
   linkClickIntercepted(element: Element, url: string): void
@@ -40,9 +38,7 @@ export class LinkInterceptor {
       if (this.delegate.shouldInterceptLinkClick(event.target, event.detail.url)) {
         this.clickEvent.preventDefault()
         event.preventDefault()
-        
-        this.convertLinkWithMethodClickToFormSubmission(event.target) || 
-          this.delegate.linkClickIntercepted(event.target, event.detail.url)
+        this.delegate.linkClickIntercepted(event.target, event.detail.url)
       }
     }
     delete this.clickEvent
@@ -50,21 +46,6 @@ export class LinkInterceptor {
 
   willVisit = () => {
     delete this.clickEvent
-  }
-
-  convertLinkWithMethodClickToFormSubmission(link: Element) {
-    const linkMethod = link.getAttribute("data-turbo-method") || link.getAttribute("data-method")
-
-    if (linkMethod) {
-      const form = document.createElement("form")
-      form.method = linkMethod
-      form.action = link.getAttribute("href") || "undefined"
-
-      link.parentNode?.insertBefore(form, link)
-      return dispatch("submit", { target: form })
-    } else {
-      return false
-    }
   }
 
   respondsToEventTarget(target: EventTarget | null) {
